+++ projects/demo-app/src/app/app.module.ts
@@ -3,6 +3,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
 import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { BrowserModule } from '@angular/platform-browser';
@@ -40,6 +41,7 @@ const routes: Routes = [
         MatSidenavModule,
         MatToolbarModule,
         MatIconModule,
+        MatListModule,
         MatButtonModule,
         RouterModule.forRoot(routes),
     ],