+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.1.0 - 2020-12-18
+> :exclamation: **Deprecations have been added in this version**
+* Implemented formatSelected configuration callback.
+This makes the customTemplate.label obsolete. The whole type of customTemplate will probably change to TemplateRef in v2.
+
+You will be notified during development of this deprecation by console warning, and only when you are using a custom template with a label defined.
+The warning is turned off in production
+
+## 1.0.1 - 2020-12-18
+* Accepted installs from angular version 9 and above
+
 ## 0.3.2 - 2020-05-10
 
 ### Fixes